refactor(module:upload): add tests
